basic refactoring for setup
