+++ MANIFEST.in
@@ -1,0 +1,5 @@
+CHANGELOG.md
+config.sample.yaml
+LICENSE
+README.md
+src/bash/listlogs
