+++ setup.cfg
@@ -1,0 +1,51 @@
+[metadata]
+name = plotman
+version = file: VERSION
+author = Eric Altendorf
+author-email = PLACEHOLDER
+home-page = https://github.com/ericaltendorf/plotman
+description = Chia plotting manager
+long-description = file: README.md
+long_description_content_type = text/markdown
+license = Apache 2.0
+license-file = LICENSE
+platform = any
+keywords = chia, blockchain, automation, process management
+classifiers =
+    Development Status :: 3 - Alpha
+    Environment :: Console :: Curses
+    Intended Audience :: Developers
+    Intended Audience :: System Administrators
+    Intended Audience :: Information Technology
+    License :: OSI Approved :: Apache Software License
+    Natural Language :: English
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+    Topic :: Office/Business :: Scheduling
+    Topic :: System :: Monitoring
+    Topic :: System :: Systems Administration
+    Topic :: Utilities
+project_urls = 
+    Bug Tracker = https://github.com/ericaltendorf/plotman/issues
+    Changelog = https://github.com/ericaltendorf/plotman/blob/main/CHANGELOG.md
+
+[options]
+package_dir=
+    =src
+packages=find:
+install_requires =
+    psutil
+    pyyaml
+    texttable
+
+[options.packages.find]
+where=src
+
+[options.entry_points]
+console_scripts =
+    plotman = plotman.plotman:main
+
+
+[options.data_files]
+config = config.yaml
+bin = util/listlogs
