+++ src/plotman/_tests/manager_test.py
