+++ src/plotman/_tests/reporting_test.py
