+++ src/plotman/manager.py
@@ -13,9 +13,9 @@ import subprocess
 import sys
 
 # Plotman libraries
-import job
-import plot_util
-import archive # for get_archdir_freebytes(). TODO: move to avoid import loop
+from plotman import job
+from plotman import plot_util
+from plotman import archive # for get_archdir_freebytes(). TODO: move to avoid import loop
 
 # Constants
 MIN = 60    # Seconds
